% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silver_searcher.R
\name{ag}
\alias{ag}
\title{Search for a pattern in a project using \code{ag}}
\usage{
ag(search, filetype = "R")
}
\arguments{
\item{search}{The pattern to search for}

\item{filetype}{The type of file to search in. One of "R", "python", or
"markdown".}
}
\value{
Search results returned by \code{ag} system call
}
\description{
\code{ag} - The Silver Searcher is a great tool for searching within files.
}
\examples{
ag("my_pattern")
ag("my_pattern", filetype = "python")
}
