% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamestools-package.R
\docType{package}
\name{jamestools-package}
\alias{jamestools}
\alias{jamestools-package}
\title{jamestools: Convenience Tools I Like to Use}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JamesHWade/jamestools}
  \item \url{https://jameshwade.github.io/jamestools/}
  \item Report bugs at \url{https://github.com/JamesHWade/jamestools/issues}
}

}
\author{
\strong{Maintainer}: James Wade \email{github@jameshwade.com} (\href{https://orcid.org/0000-0002-9740-1905}{ORCID})

}
\keyword{internal}
